Fix empty display and repeated floating point
Previously, if you clicked any button and then input an equation through
the keyboard, the output would act as if said button was always clicked.
Also, if the output of the equation were a decimal the user could enter
another floating point.
Now, the script checks if there are any floating point on the current
equation before allowing new.